change ean number to be a String as there is no need to have that as a Long

+++ composeApp/src/androidMain/kotlin/de/dias/dominik/MainActivity.kt
@@ -85,7 +85,7 @@ class MainActivity : ComponentActivity() {
                                                 scope.launch {
                                                     mainActivityViewModel
                                                         .requestProduct(
-                                                            mainActivityViewModel.barcodeText.toLong(),
+                                                            mainActivityViewModel.barcodeText,
                                                         ).collect {
                                                             url = it.products.firstOrNull()?.image_thumb_url ?: ""
                                                         }
@@ -114,7 +114,7 @@ class MainActivity : ComponentActivity() {
                                                     scope.launch {
                                                         mainActivityViewModel
                                                             .requestProduct(
-                                                                mainActivityViewModel.barcodeText.toLong(),
+                                                                mainActivityViewModel.barcodeText,
                                                             ).collect {
                                                                 url =
                                                                     it.products.firstOrNull()?.image_thumb_url
